refactor(adapters): extract Toolathlon source adapter
